Add files via upload
Model Development

Context

  - The attached notebook shows the development and visualization of several models that predict the price of a car using the variables or features within the 
    dataset. This allows for an estimate of how much a car should cost.

Following the notebook will allow for an understanding of the model types, and how appropriate they are.

Model Types Used

     -  The following model types are considered:

              - Linear Regression
              - Multiple Linear Regression
              - Polynomial Regression
              - Pipelines

There are many steps to getting a prediction. For example, normalization, polynomial transform, and linear regression. 
These processes are simplifiedusing a pipeline. 
Pipeline sequentially perform a series of transformations. 
To do so, the module Pipeline is used here to create a pipeline.